+++ packages/expo-dev-menu/CHANGELOG.md
@@ -13,6 +13,7 @@
 ### 💡 Others
 
 - Remove deprecated `REACT_NATIVE_OVERRIDE_VERSION` for React Native nightly testing. ([#25151](https://github.com/expo/expo/pull/25151) by [@kudo](https://github.com/kudo))
+- Removed backward compatible code for deprecated SDKs. ([#25154](https://github.com/expo/expo/pull/25154) by [@kudo](https://github.com/kudo))
 
 ## 4.3.0 — 2023-10-17